Moved methods from ResourceStatusSystem/CSHelpers to DMSHelpers
