+++ ResourceStatusSystem/Utilities/CSHelpers.py
@@ -6,14 +6,11 @@
 
 __RCSID__ = '$Id$'
 
-import errno
-
-from DIRAC import gConfig, gLogger, S_OK, S_ERROR
+from DIRAC import gConfig, gLogger, S_OK
 from DIRAC.Core.Utilities.Decorators import deprecated
 from DIRAC.ConfigurationSystem.Client.Helpers.Resources import getQueues
-from DIRAC.DataManagementSystem.Utilities.DMSHelpers import DMSHelpers
+from DIRAC.DataManagementSystem.Utilities.DMSHelpers import DMSHelpers, getSEParameters
 from DIRAC.ResourceStatusSystem.Utilities import Utils
-from DIRAC.Resources.Storage.StorageElement import StorageElement
 
 
 def warmUp():
@@ -129,61 +126,6 @@ def getStorageElements():
   return seNames
 
 
-def getStorageElementsHosts(seNames=None, plugins=None):
-  """ Get StorageElement host names
-
-      :param list seNames: possible list of storage element names (if not provided, will use all)
-      :param list plugins: if provided, restrict to a certain list of plugins
-
-      :return: S_OK() with list of hosts or S_ERROR
-  """
-
-  seHosts = []
-
-  if seNames is None:
-    seNames = DMSHelpers().getStorageElements()
-
-  for seName in seNames:
-
-    seHost = getSEHost(seName, plugins)
-    if not seHost['OK']:
-      gLogger.warn("Could not get SE Host", "SE: %s" % seName)
-      continue
-    if seHost['Value']:
-      seHosts.append(seHost['Value'])
-
-  return S_OK(list(set(seHosts)))
-
-
-def getSEParameters(seName, plugins=None):
-  """ get all the SE parameters in a list
-
-      :param str seName: name of the Storage Element
-      :param list plugins: if provided, restrict to a certain list of plugins
-
-      :return: S_OK() or S_ERROR
-  """
-  se = StorageElement(seName, hideExceptions=True)
-
-  seParameters = S_ERROR(errno.ENODATA, 'No SE parameters obtained')
-  pluginsList = se.getPlugins()
-  if not pluginsList['OK']:
-    gLogger.warn(pluginsList['Message'], "SE: %s" % seName)
-    return pluginsList
-  if plugins:
-    pluginsSet = set(pluginsList['Value']).intersection(set(plugins))
-  else:
-    pluginsSet = set(pluginsList['Value'])
-
-  seParametersList = []
-  for plugin in pluginsSet:
-    seParameters = se.getStorageParameters(plugin)
-    if seParameters['OK']:
-      seParametersList.append(seParameters['Value'])
-
-  return S_OK(seParametersList)
-
-
 @deprecated("unused")
 def getSEToken(seName):
   """ Get StorageElement token
@@ -195,23 +137,6 @@ def getSEToken(seName):
     return seParameters
 
   return S_OK(seParameters['Value']['SpaceToken'])
-
-
-def getSEHost(seName, plugins=None):
-  """ Get StorageElement host names (can be more than one depending on the protocol)
-
-      :param str seName: name of the storage element
-      :param list plugins: if provided, restrict to a certain list of plugins
-
-      :return: S_OK() with list of hosts or S_ERROR
-  """
-
-  seParameters = getSEParameters(seName, plugins)
-  if not seParameters['OK']:
-    gLogger.warn("Could not get SE parameters", "SE: %s" % seName)
-    return seParameters
-
-  return S_OK([parameters['Host'] for parameters in seParameters['Value']])
 
 
 def getStorageElementEndpoint(seName, plugins=[]):
